requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
yfinance>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
